Start splitting gregory provider data (#1137)
* Start splitting gregory provider data

* Fix imports from symbols for provider
